remove --flat-playlist; add --ignore-errors + --socket-timeout 30 to fix hang + get dates
YouTube's playlist API doesn't expose exact dates in flat mode (only relative
strings like "2 years ago"), so --flat-playlist can never populate upload_date.

Remove --flat-playlist so yt-dlp fetches full per-video metadata (includes
upload_date). Add --ignore-errors to skip members-only videos and
--socket-timeout 30 to kill any video fetch that genuinely hangs.

Also fix Timestamp deserialization: yt-dlp can emit it as a JSON float,
so change the field type from long? to double? and cast in ParseTimestamp.

+++ src/Streamarr.Core/Download/YtDlp/YtDlpVideoInfo.cs
@@ -22,9 +22,9 @@ namespace Streamarr.Core.Download.YtDlp
         [JsonPropertyName("upload_date")]
         public string UploadDate { get; set; } = string.Empty;
 
-        // Unix timestamp — present in flat playlist entries even when upload_date is absent
+        // Unix timestamp — yt-dlp may emit this as a JSON float, so use double?
         [JsonPropertyName("timestamp")]
-        public long? Timestamp { get; set; }
+        public double? Timestamp { get; set; }
 
         [JsonPropertyName("channel")]
         public string Channel { get; set; } = string.Empty;
